Compile fix for FreeBSD
svn path=/trunk/extragear/network/ktorrent/; revision=621387

+++ libktorrent/net/speed.cpp
@@ -26,7 +26,7 @@ using namespace bt;
 
 namespace net
 {
-	const Uint32 SPEED_INTERVAL = 5000;
+	const Uint64 SPEED_INTERVAL = 5000;
 
 	Speed::Speed() : rate(0),bytes(0)
 	{}
